Add ApiGame as a handler. (#2804)
